+++ bin/urltest.php
@@ -25,7 +25,7 @@ function requireAutoload(): void
         }
     }
     $autoloads = $autoloads ?? [
-        __DIR__ . '/vendor/autoload.php',
+        __DIR__ . '/../vendor/autoload.php',
         __DIR__ . '/../../autoload.php',
         __DIR__ . '/../../../autoload.php'
     ];